Moved old team's code to a separate file
